pyproject.toml
[project]
name = "trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "lightgbm>=4.6.0",
    "line-bot-sdk>=3.21.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
